requirements-dev.txt
-r requirements.txt
envdir==1.0.1
black==22.8.0
isort==5.10.1
mypy==0.971
django-stubs==1.12.0